Draw_Mirror_Code
This is the code created for the DRAW instance of the Mirror object. It was used to detect when a player is in view of the mirror. We were going to eventually have the player be killed shortly after being in a Mirror's FOV for too long, but never got there in time.